Extract Tags from BCheck.
