+++ src/main/java/bcheck/BCheck.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Objects;
 
 import static java.nio.file.Files.readString;
-import static java.util.Collections.unmodifiableList;
 
 public class BCheck implements Item {
     private final String name;
@@ -81,34 +80,5 @@ public class BCheck implements Item {
     @Override
     public int hashCode() {
         return Objects.hash(name, description, author, tags, path, filename, script);
-    }
-
-    public static class Tags {
-        private final List<String> tags;
-
-        private Tags(List<String> tags) {
-            this.tags = tags;
-        }
-
-        public List<String> tags() {
-            return unmodifiableList(tags);
-        }
-
-        public boolean contains(String searchText) {
-            return tags.stream().anyMatch(tag -> tag.equalsIgnoreCase(searchText));
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-            Tags tags1 = (Tags) o;
-            return Objects.equals(tags, tags1.tags);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(tags);
-        }
     }
 }
